+++ decidim-comments/spec/events/decidim/comments/comment_created_event_spec.rb
@@ -5,38 +5,15 @@ require "spec_helper"
 describe Decidim::Comments::CommentCreatedEvent do
   include_context "when it is a comment event"
   let(:event_name) { "decidim.events.comments.comment_created" }
+  let(:email_subject) { "There is a new comment from #{comment_author.name} in #{resource_title}" }
+  let(:email_intro) { "#{resource_title} has been commented. You can read the comment in this page:" }
+  let(:email_outro) { "You have received this notification because you are following \"#{resource_title}\" or its author. You can unfollow it from the previous link." }
+  let(:notification_title) { "There is a new comment from <a href=\"/profiles/#{comment_author.nickname}\">#{comment_author.name} @#{comment_author.nickname}</a> in <a href=\"#{resource_path}?commentId=#{comment.id}#comment_#{comment.id}\">#{resource_title}</a>" }
 
+  it_behaves_like "a simple event email"
+  it_behaves_like "a simple event notification"
   it_behaves_like "a comment event"
   it_behaves_like "a translated comment event" do
     let(:translatable) { true }
-  end
-
-  describe "email_subject" do
-    it "is generated correctly" do
-      expect(subject.email_subject).to eq("There is a new comment from #{comment_author.name} in #{resource_title}")
-    end
-  end
-
-  describe "email_intro" do
-    it "is generated correctly" do
-      expect(subject.email_intro).to eq("#{resource_title} has been commented. You can read the comment in this page:")
-    end
-  end
-
-  describe "email_outro" do
-    it "is generated correctly" do
-      expect(subject.email_outro)
-        .to eq("You have received this notification because you are following \"#{resource_title}\" or its author. You can unfollow it from the previous link.")
-    end
-  end
-
-  describe "notification_title" do
-    it "is generated correctly" do
-      expect(subject.notification_title)
-        .to include("There is a new comment from <a href=\"/profiles/#{comment_author.nickname}\">#{comment_author.name} @#{comment_author.nickname}</a>")
-
-      expect(subject.notification_title)
-        .to include(" in <a href=\"#{resource_path}?commentId=#{comment.id}#comment_#{comment.id}\">#{resource_title}</a>")
-    end
   end
 end