+++ decidim-conferences/app/controllers/decidim/conferences/admin/registration_types_controller.rb
@@ -10,6 +10,8 @@ module Decidim
         include Decidim::ApplicationHelper
         include Decidim::Paginable
 
+        add_breadcrumb_item_from_menu :conferences_admin_registrations_menu
+
         def index
           enforce_permission_to :index, :registration_type
 
@@ -25,7 +27,7 @@ module Decidim
           enforce_permission_to :create, :registration_type
           @form = form(Decidim::Conferences::Admin::RegistrationTypeForm).from_params(params)
 
-          CreateRegistrationType.call(@form, current_user, current_conference) do
+          CreateRegistrationType.call(@form) do
             on(:ok) do
               flash[:notice] = I18n.t("registration_types.create.success", scope: "decidim.admin")
               redirect_to conference_registration_types_path(current_conference)