+++ decidim-conferences/spec/events/decidim/conferences/conference_role_assigned_event_spec.rb
@@ -9,32 +9,12 @@ describe Decidim::Conferences::ConferenceRoleAssignedEvent do
   let(:event_name) { "decidim.events.conferences.role_assigned" }
   let(:role) { create(:conference_user_role, user:, conference: resource, role: :admin) }
   let(:extra) { { role: } }
+  let(:email_subject) { "You have been assigned as #{role} for \"#{resource.title["en"]}\"." }
+  let(:email_outro) { "You have received this notification because you are #{role} of the \"#{resource.title["en"]}\" conference." }
+  let(:email_intro) { "You have been assigned as #{role} for conference \"#{resource.title["en"]}\"." }
+  let(:notification_title) { "You have been assigned as #{role} for conference <a href=\"#{resource_url}\">#{resource.title["en"]}</a>." }
 
   it_behaves_like "a simple event"
-
-  describe "email_subject" do
-    it "is generated correctly" do
-      expect(subject.email_subject).to eq("You have been assigned as #{role} for \"#{resource.title["en"]}\".")
-    end
-  end
-
-  describe "email_outro" do
-    it "is generated correctly" do
-      expect(subject.email_outro)
-        .to eq("You have received this notification because you are #{role} of the \"#{resource.title["en"]}\" conference.")
-    end
-  end
-
-  describe "email_intro" do
-    it "is generated correctly" do
-      expect(subject.email_intro)
-        .to eq("You have been assigned as #{role} for conference \"#{resource.title["en"]}\".")
-    end
-  end
-
-  describe "notification_title" do
-    it "is generated correctly" do
-      expect(subject.notification_title).to include("You have been assigned as #{role} for conference <a href=\"#{resource_url}\">#{resource.title["en"]}</a>.")
-    end
-  end
+  it_behaves_like "a simple event email"
+  it_behaves_like "a simple event notification"
 end