+++ decidim-core/app/packs/src/decidim/i18n.test.js
@@ -1,4 +1,4 @@
-import * as i18n from "./i18n";
+import * as i18n from "src/decidim/i18n";
 
 const config = {
   messages: {