+++ decidim-core/lib/decidim/core/engine.rb
@@ -13,7 +13,6 @@ require "devise-i18n"
 require "devise_invitable"
 require "foundation_rails_helper"
 require "active_link_to"
-require "carrierwave"
 require "rails-i18n"
 require "date_validator"
 require "file_validators"
@@ -379,9 +378,6 @@ module Decidim
 
       initializer "decidim_core.menu" do
         Decidim::Core::Menu.register_menu!
-      end
-
-      initializer "decidim_core.user_menu" do
         Decidim::Core::Menu.register_user_menu!
       end