+++ decidim-proposals/app/helpers/decidim/proposals/admin/proposals_helper.rb
@@ -32,6 +32,7 @@ module Decidim
         end
 
         def proposal_complete_state(proposal)
+          return humanize_proposal_state(:withdrawn).html_safe if proposal.withdrawn?
           return humanize_proposal_state(proposal.internal_state).html_safe if proposal.answered? && !proposal.published_state?
 
           humanize_proposal_state(proposal.state).html_safe